+++ include/utils.h
@@ -1,7 +1,7 @@
 #ifndef UTILS_H
 #define UTILS_H
 
-#include <cstdint> 	/* uint32_t */
+#include <cstdint> 	/* int64_t, uint64_t */
 #include <cstdlib>	/* RAND_MAX */
 
 /**
@@ -13,7 +13,7 @@
  * @param debug	Whether to print debug information (will run entire vector)
  * @return 	Whether the two vectors are element-wise equivalent
  */
-bool compVec(uint32_t *vec1, uint32_t *vec2, uint32_t n, bool debug=false);
+bool compVec(uint64_t *vec1, uint64_t *vec2, uint64_t n, bool debug=false);
 
 /**
  * Perform the operation 'base (mod m)'
@@ -22,7 +22,7 @@ bool compVec(uint32_t *vec1, uint32_t *vec2, uint32_t n, bool debug=false);
  * @param m	The modulus of the expression
  * @return 	The result of the expression
  */
-uint32_t modulo(int32_t base, int32_t m);
+uint64_t modulo(int64_t base, int64_t m);
 
 /**
  * Perform the operation 'base^exp (mod m)' using the memory-efficient method
@@ -32,7 +32,7 @@ uint32_t modulo(int32_t base, int32_t m);
  * @param m	The modulus of the expression
  * @return 	The result of the expression
  */
-uint32_t modExp(uint32_t base, uint32_t exp, uint32_t m);
+uint64_t modExp(uint64_t base, uint64_t exp, uint64_t m);
 
 
 /**
@@ -41,7 +41,7 @@ uint32_t modExp(uint32_t base, uint32_t exp, uint32_t m);
  * @param vec	The array to be displayed
  * @param n	The length of the array
  */
-void printVec(uint32_t *vec, uint32_t n);
+void printVec(uint64_t *vec, uint64_t n);
 
 /**
  * Generate an array of arbitrary length containing random positive integers 
@@ -49,6 +49,6 @@ void printVec(uint32_t *vec, uint32_t n);
  * @param n	The length of the array
  * @param max	The maximum value for an array element [Default: RAND_MAX]
  */
-uint32_t *randVec(uint32_t n, uint32_t max=RAND_MAX);
+uint64_t *randVec(uint64_t n, uint64_t max=RAND_MAX);
 
 #endif
